requirements.txt
Django==3.1.*
Pillow==8.0.*
django-admin-sortable2==0.7.8
django-tinymce==3.2.0
python-dotenv==0.15.0
requests==2.25.1